\name{age.profile}
\alias{age.profile}
\title{Calculate age-specific population counts}

\description{
Not intended to be called directly.  
} 

\usage{
age.profile(ped, sex, year)
}

\arguments{
\item{ped}{an object returned by the \code{\link[poplab]{create.baseline.complete}}
or the \code{\link[poplab]{simped}} function}
\item{sex}{1/2 values - gender specification (1 = males, 2 = females)}
\item{year}{year for which to calculate the age-specific counts}
}
\keyword{internal}

