\name{assigndeath}
\alias{assigndeath}
\title{Assign death events}

\description{
Not intended to be called directly.  
} 

\usage{
assigndeath(ped, tempmort, risk, yr) 
}

\arguments{
\item{ped}{an object storing pedigree information}
\item{tempmort}{an object specifying age- and gender-specific mortality rates for the simulated year}
\item{risk}{multiplication value of the population mortality rates for diseased (or condition of interest) cases}
\item{yr}{year for which to assign death events}
}

\keyword{internal}
