\name{decide}
\alias{decide}
\title{"Tailor" the simulated population to match the real population age-profile}

\description{
Not intended to be called directly.  
} 

\usage{
decide(ped, sex, pop, baseyr)
}

\arguments{
\item{ped}{an object storing pedigree information}
\item{sex}{1/2 values - gender specification (1 = males, 2 = females)}
\item{pop}{real population counts for the baseline year}
\item{baseyr}{baseline year}
}

\keyword{internal}