\name{draw.profile}
\alias{draw.profile}
\title{Plot gender-specific age-profile of the real and simulated populations}

\description{
Not intended to be called directly.  
} 

\usage{
draw.profile(pop, year, realmale, realfem) 
}

\arguments{
\item{pop}{an object storing pedigree information (the simulated population)}
\item{year}{year for which to display the age-profile}
\item{realmale}{real male population counts for the considered year}
\item{realfem}{real female population counts for the considered year}
}

\keyword{internal}
