\name{givebirth}
\alias{givebirth}
\title{Create new individuals (births) in the population}

\description{
Not intended to be called directly.  
} 

\usage{
givebirth(ped, childbc, nr, yr)
}

\arguments{
\item{ped}{an object storing pedigree information}
\item{childbc}{an object specifying age-specific fertility rates for the 
simulated year}
\item{nr}{maximum value of the ID number created up to (and not including) the 
simulated year}
\item{yr}{year for which to create new individuals}
}

\keyword{internal}
