\name{readin}
\alias{readin}
\title{Read-in a file from a specified folder and assign header information}

\description{
Not intended to be called directly.  
} 

\usage{
readin(f.name, folder) 
}

\arguments{
\item{f.name}{file name (cited) to be read}
\item{folder}{path (cited) from where to read the file}
}

\keyword{internal}
