\name{retrieve.parents}
\alias{retrieve.parents}
\title{Select parents of individuals from the trimmed population who were 
excluded by the trimming process}

\description{
Not intended to be called directly.  
} 

\usage{
retrieve.parents(pop, bigpop) 
}

\arguments{
\item{pop}{an object storing the trimmed population for the baseline year}
\item{bigpop}{an object storing the simulated population before trimming it}
}

\keyword{internal}
