\name{simpop.Rdata}
\docType{data}
\alias{simpop.Rdata}
\title{A simulated population}
\description{
Example of a simulated population for Sweden for the calendar period 1955-2002.
}
\usage{simpop.Rdata}
\format{list with 2 components: the baseline population and the evolved population.}
\keyword{datasets}
